[TASK] Resolve the `SharedPluginConfiguration` trait
Fixes #4968
